Added podcast view Change datesource to memories collection changed classes names
