+++ src/Calendar/NewCalendar.js
@@ -1,6 +1,6 @@
 import "./styles.css";
 import { useState } from "react";
-import Calendar, { CalendarDayHeader } from "./calendar.js";
+import Calendar, { CalendarDayHeader } from "./Calendar.js";
 
 export default function NewCalendar() {
   const [yearAndMonth, setYearAndMonth] = useState([2021, 11]);
